@@ -3,33 +3,50 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "@types/invariant": "^2.2.34",
     "@types/jest": "^26.0.15",
     "@types/lodash": "^4.14.165",
     "@types/node": "^12.19.2",
-    "@types/react": "^16.9.54",
-    "@types/react-dom": "^16.9.9",
-    "@types/react-router-dom": "^5.1.6",
+    "@types/react": "18.2.45",
+    "@types/react-dom": "18.2.18",
+    "@types/react-router-dom": "5.3.3",
     "@types/styled-components": "^5.1.4",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-google": "^0.14.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-config-react": "^1.1.7",
-    "typescript": "^4.0.5"
+    "@types/uuid": "^9.0.7",
+    "@typescript-eslint/eslint-plugin": "^6.15.0",
+    "@typescript-eslint/parser": "6.15.0",
+    "eslint": "8.48.0",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-import": "2.28.1",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-react": "7.33.2",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "prettier": "3.1.1",
+    "typescript": "5.3.3"
   },
   "dependencies": {
-    "axios": "^0.21.1",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.1",
+    "@mui/lab": "^5.0.0-alpha.157",
+    "@mui/material": "^5.15.1",
+    "axios": "0.27.2",
     "date-fns": "^2.16.1",
+    "deep-object-diff": "^1.1.9",
     "firebase": "^8.0.0",
-    "invariant": "^2.2.4",
-    "lodash": "^4.17.20",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.0",
-    "react-toastify": "^6.0.9",
+    "i18next": "^23.7.11",
+    "i18next-http-backend": "^2.4.2",
+    "lodash": "4.17.21",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-i18next": "^14.0.0",
+    "react-router-dom": "6.21.0",
+    "react-scripts": "5.0.1",
+    "react-toastify": "9.1.3",
     "styled-components": "^5.2.0",
-    "styled-normalize": "^8.0.7"
+    "styled-normalize": "^8.0.7",
+    "uuid": "^9.0.1",
+    "zustand": "^4.4.7"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -58,5 +75,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "engines": {
+    "node": "18.x"
   }
 }
